refac: ollama gguf upload
